+++ src/sagemaker/huggingface/model.py
@@ -293,10 +293,6 @@ class HuggingFaceModel(FrameworkModel):
             str: The appropriate image URI based on the given parameters.
 
         """
-        if image_uris._processor(instance_type, ["cpu", "gpu"]) == "gpu":
-            container_version = "cu110-ubuntu18.04"
-        else:
-            container_version = "ubuntu18.04"
         if self.tensorflow_version is not None:  # pylint: disable=no-member
             base_framework_version = (
                 f"tensorflow{self.tensorflow_version}"  # pylint: disable=no-member
@@ -312,5 +308,4 @@ class HuggingFaceModel(FrameworkModel):
             accelerator_type=accelerator_type,
             image_scope="inference",
             base_framework_version=base_framework_version,
-            container_version=container_version,
         )